* Moving code out of the `__init__.py` file and into `pyro.py`. * Starting to use some `get_config` calls. * Consolidating the files responsible for starting Pyro servers.
